Replaced the sidebar component from Dashboard to the global component
