+++ frontend/src/app/components/layouts/sidebar/SidebarSubmenu.tsx
@@ -1,17 +1,16 @@
-import React, { useState } from 'react'
-import Link from 'next/link'
-import { DropdownIcon } from '../../../../icons'
-import routes from '../../routes/sidebar'
-import { Transition } from '@windmill/react-ui'
-import { Icon } from './utilities'
-import { SidebarSubmenuProps } from '../../interface'
-
+import React, { useState } from 'react';
+import Link from 'next/link';
+import { DropdownIcon } from '../../../../icons';
+import routes from '../../../../features/dashboard/routes/sidebar';
+import { Transition } from '@windmill/react-ui';
+import { Icon } from './utilities';
+import { SidebarSubmenuProps } from '../../../../features/dashboard/interface';
 
 const SidebarSubmenu = ({ route }: SidebarSubmenuProps) => {
-    const [isDropdownMenuOpen, setIsDropdownMenuOpen] = useState(false)
+    const [isDropdownMenuOpen, setIsDropdownMenuOpen] = useState(false);
 
     function handleDropdownMenuClick() {
-        setIsDropdownMenuOpen(!isDropdownMenuOpen)
+        setIsDropdownMenuOpen(!isDropdownMenuOpen);
     }
 
     return (
@@ -22,7 +21,11 @@ const SidebarSubmenu = ({ route }: SidebarSubmenuProps) => {
                 aria-haspopup="true"
             >
                 <span className="inline-flex items-center">
-                    <Icon className="w-5 h-5" aria-hidden="true" icon={route.icon} />
+                    <Icon
+                        className="w-5 h-5"
+                        aria-hidden="true"
+                        icon={route.icon}
+                    />
                     <span className="ml-4">{route.name}</span>
                 </span>
                 {/* <DropdownIcon className="w-4 h-4" aria-hidden="true" /> */}
@@ -40,20 +43,21 @@ const SidebarSubmenu = ({ route }: SidebarSubmenuProps) => {
                     className="p-2 mt-2 space-y-2 overflow-hidden text-sm font-medium text-gray-500 rounded-md shadow-inner bg-gray-50 dark:text-gray-400 dark:bg-gray-900"
                     aria-label="submenu"
                 >
-                    {route.routes && route.routes.map((r) => (
-                        <li
-                            className="px-2 py-1 transition-colors duration-150 hover:text-gray-800 dark:hover:text-gray-200"
-                            key={r.name}
-                        >
-                            <Link className="w-full" href={r.path}>
-                                {r.name}
-                            </Link>
-                        </li>
-                    ))}
+                    {route.routes &&
+                        route.routes.map((r) => (
+                            <li
+                                className="px-2 py-1 transition-colors duration-150 hover:text-gray-800 dark:hover:text-gray-200"
+                                key={r.name}
+                            >
+                                <Link className="w-full" href={r.path}>
+                                    {r.name}
+                                </Link>
+                            </li>
+                        ))}
                 </ul>
             </Transition>
         </li>
-    )
-}
+    );
+};
 
-export default SidebarSubmenu
+export default SidebarSubmenu;
